refactor product service
